refactor(renderer/api/miniGame): move interfaces to co-located types/ folder
