+++ src/renderer/api/miniGame/types/index.ts
@@ -1,0 +1,31 @@
+/*
+ * eIsland - A sleek, Apple Dynamic Island inspired floating widget for Windows, built with Electron.
+ * https://github.com/JNTMTMTM/eIsland
+ *
+ * Copyright (C) 2026 JNTMTMTM
+ * Copyright (C) 2026 pyisland.com
+ *
+ * Original author: JNTMTMTM[](https://github.com/JNTMTMTM)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ */
+
+/**
+ * @file index.ts
+ * @description 小游戏模块类型统一导出
+ * @author 鸡哥
+ */
+
+export type { MiniGameScoreData } from './MiniGameScoreData';
+export type { MiniGameLeaderboardEntry } from './MiniGameLeaderboardEntry';
+export type { MiniGameSubmitPayload } from './MiniGameSubmitPayload';
+export type { MiniGameSessionData } from './MiniGameSessionData';
+export type { LeaderboardRefreshCheckData } from './LeaderboardRefreshCheckData';
